WIP: refactoring KeyEditPage
